Split code of 2xml into 2xml and 2bsm
